requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.0
python-dotenv==1.0.0
pytest==7.4.3
